Ajout de la possibilité de spécifier la durée dans URL dans le .env

+++ DataObject/src/main/java/ch/heigvd/amt/team09/dataobject/dto/DataObjectWithUrl.java
@@ -2,5 +2,5 @@ package ch.heigvd.amt.team09.dataobject.dto;
 
 import java.net.URL;
 
-public record DataObjectWithUrl(String objectName, URL url, Integer duration) implements DataObjectResponse {
+public record DataObjectWithUrl(String objectName, URL url, long duration) implements DataObjectResponse {
 }
